move initializeFormatters to own file
